Destroy Tanker in close(), since Javascript doesn't have finalizers
There needs to be _some_ way to free up the native resources
that have been allocated once they become unused, but Javascript
has no way of communicating that an object has been collected.

We automatically re-create the instance as needed in the (unlikely)
scenario that the object is reused after close().