feat: Import test suite and run each test with Detox
This imports the same tests that we had in the previous version of the
React Native SDK, these run on device in the RN app, and they use
a very basic custom test framework to redefine describe()/it().

Since Detox is not aware of our homerolled test registry,
we previously were running EVERY on-device test under a single it()
in Jest/Detox, and the whole thing was pass/fail as a single chunk.

In this commit we extend the test framework a bit to export a list
of tests in JSON, and we read this list from the Detox globalSetup
to create a 1:1 map from our custom it() to a matching Detox it().

Each test now runs separately (which allows parallelization),
and we refresh the RN app between each test for a clean environment.