fix: Workaround Detox bug on Android with New Architecture
The Detox network synchronization class tries to find the
React Native OkHttpClient by reflection, in a native java module of
React Native.

Unfortunately when the new architecture (specifically Fabric) is
enabled, Detox looks for the HTTP client _before_ it is initialized.

We disable the synchronization during the app load, then in the app
initialization we force the React Native HTTP client to load,
and finally we re-enable synchronization.
At this point the HTTP client will have been created, and the
Detox reflection nonsense will work.