fix(ios): Handles must be double instead of NSNumber for new arch
The old arch can use either NSNumber* or double, but with Codegen
we don't have a choice, it generates C++ & ObjC bindings for us
with all the functions already declared, and we have to conform.

This upside is in general there is slightly less overhead with this
way of doing things, since the double can come directly from the
JSValue* without any cast.
But when we define a function with the wrong type, the error is
inscrutable. And currently, React Native's error conversion just
crashes with a hard segfault instead of telling us the problem.