fix: When using deployed sdk-ios, use our private podspec repo
There is a placeholder Tanker 0.1.0 on the public cocoapod "trunk"
repository, and by default that's where react-native looks for Tanker.

But we upload sdk-ios on a private podspec repo, so we have to
tell react-native about it in our Podfile.
Let's also require >= 4 in our podspec to prevent accidentally
depending on 0.1.0, which creates very strange errors.